// back/ecommerce-GonzaloMoreiro/src/Users/user.interface.ts
export interface User {
  id: number;
  email: string;
  name: string;
  password: string;
  address: string;
  phone: string;
  country?: string | undefined;
  city?: string | undefined;
}

export function User(
  arg0: any,
  id: string,
  string: any,
  arg3: any,
  user: any,
  User: any,
) {
  throw new Error('Function not implemented.');
}